requirements.txt
openai
pydantic
tenacity
python-dotenv
pandas
matplotlib
seaborn